requirements.txt
ansicolors>=1.0.2
contexttimer>=0.3.1
Django>=1.9.1,<2.0
django-allauth>=0.24.1,<0.25
django-crispy-forms>=1.6.0,<2
django-debug-toolbar
django-toolbelt>=0.0.1
materiality.commons>=0.1.19
psycopg2>=2.6.1,<3.0
requests>=2.10.0,<3.0
whitenoise>=2.0.6,<3.0
